Separate create importer and edit importer components
