+++ ui-react/src/components/accounts/EditImporter.module.css
@@ -1,0 +1,24 @@
+.container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.card {
+    align-items: center;
+}
+
+.logo {
+    margin-top: 10px;
+    height: 100px;
+    width: 100px;
+}
+
+.actionButtonsWrapper {
+    display: flex;
+    justify-content: space-between;
+}
+
+.formGroup {
+    margin-bottom: 10px;
+}
